requirements.txt
# Core OpenAI and API Dependencies
openai>=1.3.0
python-dotenv>=0.19.0
requests>=2.31.0
rich>=13.0.0
tenacity>=8.0.0
typing-extensions>=4.0.0

# Google API Dependencies
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pytz>=2024.1
